+++ packages/components-next/src/components/keyboard-shortcut/index.ts
@@ -1,2 +1,3 @@
 export type { KeyboardShortcutProps as VcKeyboardShortcutProps } from '@/components/keyboard-shortcut/keyboardShortcut.props'
+export { createKeyboardShortcutStyle } from '@/components/keyboard-shortcut/keyboardShortcut.style'
 export { default as VcKeyboardShortcut } from '@/components/keyboard-shortcut/KeyboardShortcut.vue'