+++ packages/components-next/src/components/table-next/parts/TableBody.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import Subgrid from '@/components/table/parts/Subgrid.vue'
+</script>
+
+<template>
+  <Subgrid role="rowgroup">
+    <slot />
+  </SubGrid>
+</template>