+++ packages/components-next/src/components/table/parts/TableBody.vue
@@ -15,5 +15,5 @@ const {
       :key="rowIndex"
       :data="rowData"
     />
-  </SubGrid>
+  </Subgrid>
 </template>