+++ packages/components-next/src/components/toast/index.ts
@@ -1,2 +1,3 @@
 export { useToast as useVcToast } from '@/components/toast/toast.composable'
+export { createToastStyle } from '@/components/toast/toast.style'
 export { default as VcToastContainer } from '@/components/toast/ToastContainer.vue'